Permission updates

See merge request seng302-2022/team-100!65

+++ portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/controller/DeadlineLifetimeController.java
@@ -2,8 +2,14 @@ package nz.ac.canterbury.seng302.portfolio.controller;
 
 import nz.ac.canterbury.seng302.portfolio.model.Deadline;
 import nz.ac.canterbury.seng302.portfolio.service.DeadlineService;
+import nz.ac.canterbury.seng302.portfolio.service.ElementService;
+import nz.ac.canterbury.seng302.portfolio.service.PermissionService;
+import nz.ac.canterbury.seng302.portfolio.service.UserAccountClientService;
+import nz.ac.canterbury.seng302.shared.identityprovider.AuthState;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -21,13 +27,30 @@ public class DeadlineLifetimeController {
     @Autowired
     private DeadlineService deadlineService;
 
+    @Autowired
+    private PermissionService permissionService;
+
+    @Autowired
+    private UserAccountClientService userAccountClientService;
+
+    @Autowired
+    private ElementService elementService;
+
     /**
      * Saves the given deadline to the database and redirects the user to the details page.
      * @param deadline new deadline to be saved
      */
     @PostMapping("/add-deadline")
-    public String deadlineSave(@ModelAttribute("deadline") Deadline deadline) {
-        deadlineService.addDeadline(deadline);
+    public String deadlineSave(@ModelAttribute("deadline") Deadline deadline,
+                               @AuthenticationPrincipal AuthState principal,
+                               Model model
+    ) {
+        Integer userID = userAccountClientService.getUserIDFromAuthState(principal);
+        elementService.addHeaderAttributes(model, userID);
+
+        if (permissionService.isValidToModifyProjectPage(userID)) {
+            deadlineService.addDeadline(deadline);
+        }
         return "redirect:/details";
     }
 
@@ -36,8 +59,15 @@ public class DeadlineLifetimeController {
      * @param id of the deadline being deleted
      */
     @GetMapping("/delete-deadline/{id}")
-    public String deadlineRemove(@PathVariable("id") Integer id) {
-        deadlineService.removeDeadline(id);
+    public String deadlineRemove(@PathVariable("id") Integer id,
+                                 @AuthenticationPrincipal AuthState principal,
+                                 Model model) {
+        Integer userID = userAccountClientService.getUserIDFromAuthState(principal);
+        elementService.addHeaderAttributes(model, userID);
+
+        if (permissionService.isValidToModifyProjectPage(userID)) {
+            deadlineService.removeDeadline(id);
+        }
         return "redirect:/details";
     }
 }
